+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 
 ### Dependency Upgrades
 
+- Upgrade to `eclipse-temurin:11.0.21_9-jre-focal`. (#121)
 - Upgrade to Spring Boot 2.7.17. (#120)
 - Upgrade to Spring Dependency Management Plugin 1.1.4. (#120)
 - Upgrade to `jenkins-library` 2.7.4. (#116)