+++ CHANGELOG.md
@@ -23,6 +23,7 @@ All notable changes to this project will be documented in this file.
 - Standardisation of the dockerfile and its location in regard to other java components. (#115)
 - Rename `Status` to `CommandStatus` in library. (#117)
 - Remove `com.iexec.blockchain.command.task.contribute` package. (#124)
+- Remove `com.iexec.blockchain.command.task.reveal` package. (#125)
 
 ### Dependency Upgrades