+++ CHANGELOG.md
@@ -32,6 +32,8 @@ All notable changes to this project will be documented in this file.
 - Upgrade to Spring Dependency Management Plugin 1.1.4. (#120)
 - Upgrade to Spring Doc OpenAPI 1.7.0. (#122)
 - Upgrade to `jenkins-library` 2.7.4. (#116)
+- Upgrade to `iexec-commons-poco` 3.2.0. (#127)
+- Upgrade to `iexec-common` 8.3.1. (#127)
 
 ## [[8.2.0]](https://github.com/iExecBlockchainComputing/iexec-blockchain-adapter-api/releases/tag/v8.2.0) 2023-09-28